time should be 64-bit its from 1970

+++ bites/util.c
@@ -18,18 +18,18 @@ double distance(double x1, double y1, double x2, double y2)
     return (x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2);
 }
 
-long msec()
+long long msec()
 {
   struct timeval tv;
   gettimeofday(&tv, 0);
   return 1000L * tv.tv_sec + tv.tv_usec / 1000L;
 }
 
-long usec()
+long long usec()
 {
   struct timeval tv;
   gettimeofday(&tv, 0);
-  return 1000000L * tv.tv_sec + tv.tv_usec;
+  return (1000000L * (long long)tv.tv_sec) + tv.tv_usec;
 }
 
 void say(char *sentence)
